source/requirements.txt
blinker==1.4
bcrypt==3.2.0
Flask==1.1.2
mongoengine==0.21.0
PyJWT==1.7.1